requirements.txt
Pillow == 9.4.0
plyer == 2.1.0
pystray == 0.19.5
pyautogui == 0.9.54
pyinstaller == 5.13.2